% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mse.mkl}
\alias{mse.mkl}
\title{Compute mean square error(MSE) and mean kL divergence (MKL)}
\usage{
mse.mkl(obs, pred, na.rm = TRUE, show.warning = TRUE)
}
\arguments{
\item{obs}{observed value}

\item{pred}{prediction/estimate}

\item{na.rm}{if to remove NAs}

\item{show.warning}{if to show warning if any}
}
\value{
A vector of c(MSE, MKL)
}
\description{
Compute mean square error(MSE) and mean kL divergence (MKL)
}
