% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{NNLM}
\alias{NNLM}
\alias{NNLM-package}
\title{Fast NNLS and NNMF}
\description{
This package is built for fast non-negative least square (NNLS) regression and non-negative matrix factorization (NNMF).
}

