% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nnls (oll-elin's conflicted copy 2015-10-30).R, R/nnls.R
\name{nnls}
\alias{nnls}
\title{Non-negative least square}
\usage{
nnls(x, y, max.iter = 500L, tol = 1e-06)

nnls(x, y, max.iter = 500L, tol = 1e-06)
}
\arguments{
\item{x}{Design mamtrix}

\item{y}{Vector or matrix of response}

\item{max.iter}{Maximum number of iterations}

\item{tol}{Stop criterion, minimum change on x between two successive iteration.}

\item{x}{Design mamtrix}

\item{y}{Vector or matrix of response}

\item{max.iter}{Maximum number of iterations}

\item{tol}{Stop criterion, minimum change on x between two successive iteration.}
}
\value{
A vector of non-negative coefficients, a solution to
	argmin_{beta} ||y - x*beta||_F^2, s.t. beta >= 0

A vector of non-negative coefficients, a solution to
	argmin_{beta} ||y - x*beta||_F^2, s.t. beta >= 0
}
\description{
Sequential Coordinate-wise algorithm for non-negative least square regression problem

Sequential Coordinate-wise algorithm for non-negative least square regression problem
}
\examples{
x <- 10*matrix(runif(50*20), 50, 20)-1
beta <- matrix(rexp(20*2), 20, 2)
y <- x \%*\% beta + matrix(rnorm(50*2), 50, 2)
beta.hat <- nnls(x, y)
x <- 10*matrix(runif(50*20), 50, 20)-1
beta <- matrix(rexp(20*2), 20, 2)
y <- x \%*\% beta + matrix(rnorm(50*2), 50, 2)
beta.hat <- nnls(x, y)
}
\author{
Eric Xihui Lin, \email{xihuil.silence@gmail.com}

Eric Xihui Lin, \email{xihuil.silence@gmail.com}
}
\references{
http://cmp.felk.cvut.cz/ftp/articles/franc/Franc-TR-2005-06.pdf

Franc, V. C.; Hlaváč, V. C.; Navara, M. (2005). Sequential Coordinate-Wise Algorithm for the Non-negative Least Squares Problem.
	Proc. Int'l Conf. Computer Analysis of Images and Patterns. Lecture Notes in Computer Science 3691. p. 407. \cr
	http://cmp.felk.cvut.cz/ftp/articles/franc/Franc-TR-2005-06.pdf
}

