% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{data}
\name{nsclc}
\alias{nsclc}
\title{Micro-array data of NSCLC patients}
\description{
This dataset is a random subset (matrix) of micro-array data from a group of Non-Small Cell Lung Caner (NSCLC) patients.
It contains 200 probes / genes (row) for 100 patients / samples (column).
}
\keyword{data}
