% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nnmf (oll-elin's conflicted copy 2015-10-30).R, R/nnmf.R
\name{predict.NNMF}
\alias{predict.NNMF}
\title{Calculate W or H matrix from a NNMF object given new matrix and pre-computed H or W}
\usage{
\method{predict}{NNMF}(object, newdata, which.matrix = c("H", "W"),
  max.iter = 500L, tol = object$tol, ...)

\method{predict}{NNMF}(object, newdata, which.matrix = c("H", "W"),
  max.iter = 500L, tol = object$tol, ...)
}
\arguments{
\item{object}{An NNMF object returned by \code{\link{nnmf}}}

\item{newdata}{A new matrix of x}

\item{which.matrix}{Either 'W' or 'H'}

\item{max.iter}{Maximum number of iterations}

\item{tol}{Stop criterion, maximum difference of target_error between two successive iterations}

\item{...}{Unsued, just for compatiblity}

\item{object}{An NNMF object returned by \code{\link{nnmf}}}

\item{newdata}{A new matrix of x}

\item{which.matrix}{Either 'W' or 'H'}

\item{max.iter}{Maximum number of iterations}

\item{tol}{Stop criterion, maximum difference of target_error between two successive iterations}

\item{...}{Unsued, just for compatiblity}
}
\value{
\code{W} or \code{H} for newdata given pre-computed H or W

\code{W} or \code{H} for newdata given pre-computed H or W
}
\description{
Calculate W or H matrix from a NNMF object given new matrix and pre-computed H or W

Calculate W or H matrix from a NNMF object given new matrix and pre-computed H or W
}
\examples{
x <- matrix(runif(50*20), 50, 20)
r <- nnmf(x, 2)
newx <- matrix(runif(30*20), 30, 20)
pred <- predict(r, newx, 'H')
x <- matrix(runif(50*20), 50, 20)
r <- nnmf(x, 2)
newx <- matrix(runif(50*30), 50, 30)
pred <- predict(r, newx, 'H')
}
\seealso{
\code{\link{nnmf}}

\code{\link{nnmf}}
}

