% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nnls.R
\name{nnls}
\alias{nnls}
\title{Non-negative least square}
\usage{
nnls(x, y, check.x = TRUE, max.iter = 10000L, tol = .Machine$double.eps)
}
\arguments{
\item{x}{Design mamtrix}

\item{y}{Vector or matrix of response}

\item{check.x}{If to check the condition number of matrix x to ensure unique solution}

\item{max.iter}{Maximum number of iterations}

\item{tol}{Stop criterion, minimum change on x between two successive iteration}
}
\value{
A vector of non-negative coefficients, a solution to
	argmin_{beta} ||y - x*beta||_F^2, s.t. beta >= 0
}
\description{
Sequential Coordinate-wise algorithm for non-negative least square regression problem
}
\examples{
x <- 10*matrix(runif(50*20), 50, 20)-1
beta <- matrix(rexp(20*2), 20, 2)
y <- x \%*\% beta + matrix(rnorm(50*2), 50, 2)
beta.hat <- nnls(x, y)
}
\author{
Eric Xihui Lin, \email{xihuil.silence@gmail.com}
}
\references{
Franc, V. C., Hlavac, V. C., Navara, M. (2005). Sequential Coordinate-Wise Algorithm for the Non-negative Least Squares Problem.
	Proc. Int'l Conf. Computer Analysis of Images and Patterns. Lecture Notes in Computer Science 3691. p. 407. \cr
	http://cmp.felk.cvut.cz/ftp/articles/franc/Franc-TR-2005-06.pdf
}

