% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnls.R
\name{nnls}
\alias{nnls}
\title{Non-negative least square (NNLS)}
\usage{
nnls(x, y, check.x = TRUE, max.iter = 10000L, rel.tol = 1e-04,
  n.threads = 1L, show.progress = TRUE)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Vector or matrix of response}

\item{check.x}{If to check the condition number of matrix x to ensure unique solution}

\item{max.iter}{Maximum number of iterations}

\item{rel.tol}{Stop criterion, relative change on x between two successive iteration}

\item{n.threads}{An integer number of threads/CPUs to use. Default to 1 (no parallel). Use 0 for all cores}

\item{show.progress}{TRUE/FALSE indicating if to show a progress bar}
}
\value{
A list of with components \itemize{
	\item{coefficients: } {a matrix or vector (depend on y) of the NNLS solution beta}
	\item{iteration: }{a vector of numbers of iterations for each column of y}
	\item{abs.err: }{a vector of absolute error, i.e., difference between two successive iterations, for each column of y}
	\item{rel.err: }{a vector of relative error, i.e., relative difference between two successive iterations, for each column of y}
	}
}
\description{
Sequential coordinate-wise algorithm for non-negative least square regression \cr
		\deqn{argmin_{\beta \ge 0} ||y - x*\beta||_F^2}
}
\examples{

x <- 10*matrix(runif(50*20), 50, 20)-1
beta <- matrix(rexp(20*2), 20, 2)
y <- x \%*\% beta + matrix(rnorm(50*2), 50, 2)
beta.hat <- nnls(x, y)

}
\author{
Eric Xihui Lin, \email{xihuil.silence@gmail.com}
}
\references{
Franc, V. C., Hlavac, V. C., Navara, M. (2005). Sequential Coordinate-Wise Algorithm for the Non-negative Least Squares Problem.
	Proc. Int'l Conf. Computer Analysis of Images and Patterns. Lecture Notes in Computer Science 3691. p. 407. \cr
	http://cmp.felk.cvut.cz/ftp/articles/franc/Franc-TR-2005-06.pdf
}

