% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmf.R
\name{nnmf}
\alias{nnmf}
\title{Non-negative matrix factorization}
\usage{
nnmf(A, k = 1L, method = c("nnls", "brunet"), W0 = NULL, H0 = NULL,
  eta = 0, beta = 0, max.iter = 500L, rel.tol = 1e-04, check.k = TRUE,
  n.threads = 1L, show.progress = TRUE, show.warning = TRUE)
}
\arguments{
\item{A}{A matrix to be decomposed}

\item{k}{An integer of decomposition rank}

\item{method}{Decomposition algorithms. Options are 'nnls'(default), 'brunet'}

\item{W0}{Known \code{W} (left) profile}

\item{H0}{Known \code{H} (right) profile}

\item{eta}{L2 penalty on the left (W). Default to no penalty. If eta < 0 then eta = max(A). Effective only when \code{method = 'nnls'}}

\item{beta}{L1 penalty on the right (H). Default to no penalty. Effective only when \code{method = 'nnls'}}

\item{max.iter}{Maximum iteration of alternating NNLS solutions to H and W}

\item{rel.tol}{Stop criterion, relative difference of target_error between two successive iterations}

\item{check.k}{If to check whether k <= n*m/(n+m), where (n,m)=dim(A)}

\item{n.threads}{An integer number of threads/CPUs to use. Default to 1(no parallel). Specify 0 for all cores}

\item{show.progress}{TRUE/FALSE indicating if to show a progress bar}

\item{show.warning}{If to show warnings when targeted \code{rel.tol} is not reached}
}
\value{
A list with components
	\itemize{
		\item{W:}{ left/base matrix W}
		\item{H:}{ right/coefficient matrix H}
		\item{error:}{ root mean square error between A and W*H}
		\item{target.error:}{ error used to stop iteration}
		\item{target.measure:}{ the measure for \code{target.error}}
		\item{H1:}{ coefficient matrix corresponding to known W0}
		\item{W1:}{ base matrix corresponding to known H0}
	}
}
\description{
Non-negative matrix factorization(NNMF) using alternating NNLS or Brunet's multiplicative update
}
\details{
The problem of non-negative matrix factorization is to find \code{W, H, W_1, H_1}, such that \cr
		\deqn{A = W H + W_0 H_1 + W_1 H_0 + noise,}\cr
where \eqn{W_0}, \eqn{H_0} are known matrices, which are NULLs in most application case. 
In tumour content deconvolution, \eqn{W_0} can be thought as known healthy profile, and \eqn{W}
is desired pure cancer profile. One also set \eqn{H_0} to a row matrix of 1, and thus \eqn{W_1}
can be treated as common profile among samples.

To simplify the notations, we denote right hand side of the above equation as \eqn{W H}. 
The problem to solved using square error is\cr
	  \deqn{argmin_{W \ge 0, H \ge 0} ||A - W H||_2^2 + \eta*||W||_F^2 + \beta*\sum_{j=1}^m (||H_j||_1^2),}\cr
where \eqn{H_j} is the j-th column of \eqn{H}. This minimization problem is solved by apply 
\code{\link{nnls}} to W and H alternatively, when \code{method == 'nnls'},
which is implemented using sequential coordinate descend methods.

When \code{method = 'brunet'}, decomposition is done using Brunet's  multiplicative updates based on 
Kullback-Leibler divergence. Note that penalties \code{eta}, \code{beta} and known profiles \code{W0} \code{H0} 
are not support in this case.
}
\examples{

x <- matrix(runif(50*20), 50, 20)
r <- nnmf(x, 2)

}
\author{
Eric Xihui Lin, \email{xihuil.silence@gmail.com}
}
\seealso{
\code{\link{nnls}}, \code{\link{predict.nnmf}}
}

