% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmf_methods.R
\name{predict.nnmf}
\alias{plot.nnmf}
\alias{predict.nnmf}
\alias{print.nnmf}
\title{Methods for nnmf object returned by \code{nnmf}}
\usage{
\method{predict}{nnmf}(object, newdata, which.matrix = c("H", "W"),
  method = object$method, max.iter = 100L, rel.tol = object$rel.tol,
  n.threads = 1L, show.progress = TRUE, show.warning = TRUE, ...)

\method{print}{nnmf}(x, ...)

\method{plot}{nnmf}(x, which = c("error", "target.error", "W", "H", "W1",
  "H1"), ...)
}
\arguments{
\item{object}{An NNMF object returned by \code{\link{nnmf}}}

\item{newdata}{A new matrix of x}

\item{which.matrix}{Either 'W' or 'H'}

\item{method}{Either 'nnls' or 'brunet'. Default to \code{object$method}}

\item{max.iter}{Maximum number of iterations}

\item{rel.tol}{Stop criterion, relative difference of target_error between two successive iterations}

\item{n.threads}{An integer number of threads/CPUs to use. Default to 1(no parallel). Use 0 for all cores}

\item{show.progress}{TRUE/FALSE indicating if to show a progress bar}

\item{show.warning}{If to show warnings when targeted \code{rel.tol} is not reached}

\item{...}{Further arguments passed to 'print', 'plot' or 'heatmap'}

\item{x}{An NNMF object returned by \code{\link{nnmf}}}

\item{which}{One of 'error', 'target', 'W', 'H', 'W1', 'H1'. The first two give line plots while the latter two give heatmaps}
}
\value{
\itemize{
	\item{predict: }{\code{W} or \code{H} for newdata given pre-computed H or W if \code{W0} and \code{H0} are NULL, else a list of \code{W, W1} or \code{H, H1}}
	\item{print: }{print running time and accuracy}
	\item{plot: }{line plot if which = 'error' or 'target.error' and heatmap if which = 'H' or 'W'}
	}
}
\description{
Methods for nnmf object returned by \code{nnmf}
}
\examples{

x <- matrix(runif(50*20), 50, 20)
r <- nnmf(x, 2)
r
newx <- matrix(runif(50*30), 50, 30)
pred <- predict(r, newx, 'H')

# check convergence
plot(r)

# show W matrix
plot(r, 'W')

}
\seealso{
\code{\link{nnmf}}
}

